+++ Sources/OpenAPIClient/Models/ConnectAppendFinishReq.swift
@@ -10,13 +10,19 @@ import Foundation
 public struct ConnectAppendFinishReq: Sendable, Codable, ParameterConvertible, Hashable {
 
     public var attestationResponse: String
+    public var completionType: AppendCompletionType
+    public var customData: [String: String]?
 
-    public init(attestationResponse: String) {
+    public init(attestationResponse: String, completionType: AppendCompletionType, customData: [String: String]? = nil) {
         self.attestationResponse = attestationResponse
+        self.completionType = completionType
+        self.customData = customData
     }
 
     public enum CodingKeys: String, CodingKey, CaseIterable {
         case attestationResponse
+        case completionType
+        case customData
     }
 
     // Encodable protocol methods
@@ -24,6 +30,8 @@ public struct ConnectAppendFinishReq: Sendable, Codable, ParameterConvertible, H
     public func encode(to encoder: Encoder) throws {
         var container = encoder.container(keyedBy: CodingKeys.self)
         try container.encode(attestationResponse, forKey: .attestationResponse)
+        try container.encode(completionType, forKey: .completionType)
+        try container.encodeIfPresent(customData, forKey: .customData)
     }
 }