+++ Sources/OpenAPIClient/Models/ConnectAppendStartRsp.swift
@@ -18,12 +18,18 @@ public struct ConnectAppendStartRsp: Sendable, Codable, ParameterConvertible, Ha
     public var variant: Variant
     public var isRestrictedBrowser: Bool
     public var autoAppend: Bool
+    public var conditionalAppend: Bool
+    public var expiresMs: Int64
+    public var customData: [String: String]?
 
-    public init(attestationOptions: String, variant: Variant, isRestrictedBrowser: Bool, autoAppend: Bool) {
+    public init(attestationOptions: String, variant: Variant, isRestrictedBrowser: Bool, autoAppend: Bool, conditionalAppend: Bool, expiresMs: Int64, customData: [String: String]? = nil) {
         self.attestationOptions = attestationOptions
         self.variant = variant
         self.isRestrictedBrowser = isRestrictedBrowser
         self.autoAppend = autoAppend
+        self.conditionalAppend = conditionalAppend
+        self.expiresMs = expiresMs
+        self.customData = customData
     }
 
     public enum CodingKeys: String, CodingKey, CaseIterable {
@@ -31,6 +37,9 @@ public struct ConnectAppendStartRsp: Sendable, Codable, ParameterConvertible, Ha
         case variant
         case isRestrictedBrowser
         case autoAppend
+        case conditionalAppend
+        case expiresMs
+        case customData
     }
 
     // Encodable protocol methods
@@ -41,6 +50,9 @@ public struct ConnectAppendStartRsp: Sendable, Codable, ParameterConvertible, Ha
         try container.encode(variant, forKey: .variant)
         try container.encode(isRestrictedBrowser, forKey: .isRestrictedBrowser)
         try container.encode(autoAppend, forKey: .autoAppend)
+        try container.encode(conditionalAppend, forKey: .conditionalAppend)
+        try container.encode(expiresMs, forKey: .expiresMs)
+        try container.encodeIfPresent(customData, forKey: .customData)
     }
 }