+++ Sources/OpenAPIClient/Models/NativeMeta.swift
@@ -26,8 +26,13 @@ public struct NativeMeta: Sendable, Codable, ParameterConvertible, Hashable {
     public var isGooglePlayServices: Bool?
     public var isDeviceSecure: Bool?
     public var error: String?
+    public var brand: String?
+    public var model: String?
+    public var locale: String?
+    public var screen: NativeMetaScreen?
+    public var sdkInitTimeMs: Int64?
 
-    public init(platform: String, platformVersion: String, name: String? = nil, version: String? = nil, displayName: String, build: String? = nil, deviceOwnerAuth: DeviceOwnerAuth? = nil, isBluetoothAvailable: Bool? = nil, isBluetoothOn: Bool? = nil, isGooglePlayServices: Bool? = nil, isDeviceSecure: Bool? = nil, error: String? = nil) {
+    public init(platform: String, platformVersion: String, name: String? = nil, version: String? = nil, displayName: String, build: String? = nil, deviceOwnerAuth: DeviceOwnerAuth? = nil, isBluetoothAvailable: Bool? = nil, isBluetoothOn: Bool? = nil, isGooglePlayServices: Bool? = nil, isDeviceSecure: Bool? = nil, error: String? = nil, brand: String? = nil, model: String? = nil, locale: String? = nil, screen: NativeMetaScreen? = nil, sdkInitTimeMs: Int64? = nil) {
         self.platform = platform
         self.platformVersion = platformVersion
         self.name = name
@@ -40,6 +45,11 @@ public struct NativeMeta: Sendable, Codable, ParameterConvertible, Hashable {
         self.isGooglePlayServices = isGooglePlayServices
         self.isDeviceSecure = isDeviceSecure
         self.error = error
+        self.brand = brand
+        self.model = model
+        self.locale = locale
+        self.screen = screen
+        self.sdkInitTimeMs = sdkInitTimeMs
     }
 
     public enum CodingKeys: String, CodingKey, CaseIterable {
@@ -55,6 +65,11 @@ public struct NativeMeta: Sendable, Codable, ParameterConvertible, Hashable {
         case isGooglePlayServices
         case isDeviceSecure
         case error
+        case brand
+        case model
+        case locale
+        case screen
+        case sdkInitTimeMs
     }
 
     // Encodable protocol methods
@@ -73,6 +88,11 @@ public struct NativeMeta: Sendable, Codable, ParameterConvertible, Hashable {
         try container.encodeIfPresent(isGooglePlayServices, forKey: .isGooglePlayServices)
         try container.encodeIfPresent(isDeviceSecure, forKey: .isDeviceSecure)
         try container.encodeIfPresent(error, forKey: .error)
+        try container.encodeIfPresent(brand, forKey: .brand)
+        try container.encodeIfPresent(model, forKey: .model)
+        try container.encodeIfPresent(locale, forKey: .locale)
+        try container.encodeIfPresent(screen, forKey: .screen)
+        try container.encodeIfPresent(sdkInitTimeMs, forKey: .sdkInitTimeMs)
     }
 }