+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.3.0] - 2025-10-01
+- Add support for situational appends
+- Add support for conditional create
+- Add support for signalAllAcceptedCredentials
+- Collect device brand
+
 ## [1.2.1] - 2025-09-30
 - Fix build error (Non-Sendable ASAuthorizationPlatformPublicKeyCredentialDescriptor in parseCredentials)